# Conflicts:
#	core/android/local.properties
#	core/lib/dto/enums/app_screen_enum.dart
#	core/lib/dto/modules/app_color_module.dart
#	core/lib/dto/network/api_client.dart
#	core/lib/dto/network/api_client.g.dart
#	core/lib/dto/network/api_client_key.dart
#	core/lib/generated/l10n.dart
#	core/lib/l10n/intl_ar.arb
#	core/lib/l10n/intl_en.arb
#	core/linux/flutter/ephemeral/.plugin_symlinks/connectivity_plus
#	core/linux/flutter/ephemeral/.plugin_symlinks/device_info_plus
#	core/linux/flutter/ephemeral/.plugin_symlinks/package_info_plus
#	core/linux/flutter/ephemeral/.plugin_symlinks/shared_preferences_linux
#	core/linux/flutter/ephemeral/.plugin_symlinks/url_launcher_linux
#	lib/my_app.dart
#	lib/navigation_routes.dart
#	more/lib/ui/more/more_widget.dart
#	pubspec.yaml

+++ more/lib/ui/more/more_widget.dart
@@ -1,15 +1,21 @@
 import 'package:core/core.dart';
 import 'package:core/dto/enums/app_screen_enum.dart';
+import 'package:core/dto/models/balance/balance_mapper.dart';
+import 'package:core/dto/models/baseModules/api_state.dart';
+import 'package:core/dto/models/profile/profile_mapper.dart';
 import 'package:core/dto/modules/alert_module.dart';
 import 'package:core/dto/modules/app_color_module.dart';
 import 'package:core/dto/modules/app_provider_module.dart';
 import 'package:core/dto/modules/custom_navigator_module.dart';
 import 'package:core/dto/modules/custom_text_style_module.dart';
+import 'package:core/dto/modules/shared_pref_module.dart';
 import 'package:core/generated/l10n.dart';
 import 'package:core/ui/bases/base_state.dart';
 import 'package:core/ui/contactUs/contact_us_bloc.dart';
+import 'package:core/ui/custom_button_widget.dart';
 import 'package:core/ui/custom_text.dart';
 import 'package:flutter/material.dart';
+import 'package:image_picker/image_picker.dart';
 import 'package:more/ui/more/more_bloc.dart';
 import 'package:more/ui/more/shop_logo_camera_widget.dart';
 
@@ -26,7 +32,6 @@ class MoreWidget extends BaseStatefulWidget {
   final String faqIcon;
   final String logoutIcon;
   final MoreBloc moreBloc;
-  final VoidCallback openCamera;
   final String usagePolicyIcon;
   final String alertIcon;
   final ContactUsBloc contactUsBloc;
@@ -45,7 +50,6 @@ class MoreWidget extends BaseStatefulWidget {
     required this.logoutIcon,
     required this.previewsOrderIcon,
     required this.shopIcon,
-    required this.openCamera,
     required this.usagePolicyIcon,
     required this.alertIcon,
     required this.contactUsBloc,
@@ -59,6 +63,7 @@ class _MoreWidgetState extends BaseState<MoreWidget> {
   @override
   void initState() {
     super.initState();
+    widget.moreBloc.getProfileData();
   }
 
   @override
@@ -71,33 +76,77 @@ class _MoreWidgetState extends BaseState<MoreWidget> {
   bool isSafeArea() => true;
 
   @override
-  Widget getBody(BuildContext context) => ListView(
-        shrinkWrap: true,
-        children: [
-          Container(
-            height: 135.h,
-            width: MediaQuery.of(context).size.width,
-            decoration: BoxDecoration(
-                color: secondaryColor,
-                borderRadius: BorderRadius.only(
-                    bottomRight: Radius.circular(20.w),
-                    bottomLeft: Radius.circular(20.w))),
-            child: _logoWidget,
+  void onPopInvoked(didPop) {
+    handleCloseApplication();
+    super.onPopInvoked(didPop);
+  }
+
+  @override
+  Widget getBody(BuildContext context) =>
+      StreamBuilder<ApiState<ProfileMapper>>(
+        stream: widget.moreBloc.userStream,
+        initialData: LoadingState(),
+        builder: (context, snapshot) => checkResponseStateWithLoadingWidget(
+            snapshot.data!, context,
+            idleWidget: _screenDesign(context, snapshot),
+            onSuccess: _screenDesign(context, snapshot)),
+      );
+
+  ListView _screenDesign(
+      BuildContext context, AsyncSnapshot<ApiState<ProfileMapper>> snapshot) {
+    return ListView(
+      shrinkWrap: true,
+      children: [
+        Container(
+          height: 135.h,
+          width: MediaQuery.of(context).size.width,
+          decoration: BoxDecoration(
+              color: secondaryColor,
+              borderRadius: BorderRadius.only(
+                  bottomRight: Radius.circular(20.w),
+                  bottomLeft: Radius.circular(20.w))),
+          child: _logoWidget,
+        ),
+        if ((SharedPrefModule().userId ?? '').isNotEmpty) ...[
+          SizedBox(
+              height: 100.h,
+              child: _imageWithCameraWidget(
+                  mobile: snapshot.data?.response?.mobile ?? '',
+                  name: snapshot.data?.response?.shopName ?? '',
+                  image: snapshot.data?.response?.image ?? '')),
+        ] else ...[
+          SizedBox(
+            height: 43.h,
           ),
-          SizedBox(height: 100.h, child: _imageWithCameraWidget),
-          // _ordersWidget,
           Padding(
-            padding: EdgeInsets.symmetric(horizontal: 16.w),
-            child: CustomText(
-                text: S.of(context).settings,
-                customTextStyle:
-                    BoldStyle(fontSize: 20.sp, color: secondaryColor)),
-          ),
+              padding: EdgeInsets.symmetric(horizontal: 14.w),
+              child: CustomButtonWidget(
+                idleText: S.of(context).loginNow,
+                onTap: () => AppProviderModule().logout(context),
+                textStyle: MediumStyle(fontSize: 16.sp, color: lightBlackColor)
+                    .getStyle(),
+                height: 60.h,
+              )),
+          SizedBox(
+            height: 27.h,
+          )
+        ],
+        // _ordersWidget,
+        Padding(
+          padding: EdgeInsets.symmetric(horizontal: 16.w),
+          child: CustomText(
+              text: S.of(context).settings,
+              customTextStyle:
+                  BoldStyle(fontSize: 20.sp, color: secondaryColor)),
+        ),
+        if ((SharedPrefModule().userId ?? '').isNotEmpty) ...[
           SizedBox(
             height: 10.h,
           ),
-          _menuItem(
-              S.of(context).accountInfo, widget.accountSettingIcon, () {}),
+          _menuItem(S.of(context).accountInfo, widget.accountSettingIcon, () {
+            CustomNavigatorModule.navigatorKey.currentState
+                ?.pushNamed(AppScreenEnum.updateProfileScreen.name);
+          }),
           SizedBox(
             height: 10.h,
           ),
@@ -106,14 +155,16 @@ class _MoreWidgetState extends BaseState<MoreWidget> {
             CustomNavigatorModule.navigatorKey.currentState
                 ?.pushNamed(AppScreenEnum.accountChangePassword.name);
           }),
+        ],
+        SizedBox(
+          height: 10.h,
+        ),
+        _menuItem(S.of(context).currentOrder, widget.myOrdersIcon, () {},
+            disabled: (SharedPrefModule().userId ?? '').isEmpty),
+        if ((SharedPrefModule().userId ?? '').isNotEmpty) ...[
           SizedBox(
             height: 10.h,
           ),
-          _menuItem(S.of(context).currentOrders, widget.myOrdersIcon, () {}),
-          SizedBox(
-            height: 10.h,
-          ),
-
           _menuItem(S.of(context).deleteAccount, widget.deleteAccountIcon, () {
             _deleteAccount();
           }),
@@ -130,7 +181,9 @@ class _MoreWidgetState extends BaseState<MoreWidget> {
           SizedBox(
             height: 18.h,
           ),
-          _accountBalance('-1190 ج.م.'),
+          _accountBalance()
+        ],
+        if ((SharedPrefModule().userId ?? '').isEmpty) ...[
           SizedBox(
             height: 18.h,
           ),
@@ -142,53 +195,135 @@ class _MoreWidgetState extends BaseState<MoreWidget> {
             ),
           ),
           SizedBox(
-            height: 10.h,
+            height: 18.h,
           ),
           Padding(
             padding: EdgeInsets.symmetric(horizontal: 16.w),
             child: CustomText(
-                text: S.of(context).supportAndAssistance,
+                text: S.of(context).accountBalance,
                 customTextStyle:
                     BoldStyle(fontSize: 20.sp, color: secondaryColor)),
+          )
+        ],
+        SizedBox(
+          height: 18.h,
+        ),
+        Padding(
+          padding: EdgeInsets.symmetric(horizontal: 16.w),
+          child: Divider(
+            height: 1.h,
+            color: secondaryColor,
           ),
-          SizedBox(
-            height: 10.h,
-          ),
-          _menuItem(S.of(context).contactUs, widget.contactUsIcon, () {
-            AlertModule().showContactUsDialog(
-                contactUsBloc: widget.contactUsBloc, context: context);
-          }),
-          SizedBox(
-            height: 10.h,
-          ),
-          _menuItem(S.of(context).faq, widget.faqIcon, () {
-            CustomNavigatorModule.navigatorKey.currentState
-                ?.pushNamed(AppScreenEnum.faq.name);
-          }),
-          SizedBox(
-            height: 10.h,
-          ),
-          _menuItem(S.of(context).usagePolicy, widget.usagePolicyIcon, () {}),
+        ),
+        SizedBox(
+          height: 10.h,
+        ),
+        Padding(
+          padding: EdgeInsets.symmetric(horizontal: 16.w),
+          child: CustomText(
+              text: S.of(context).supportAndAssistance,
+              customTextStyle:
+                  BoldStyle(fontSize: 20.sp, color: secondaryColor)),
+        ),
+        SizedBox(
+          height: 10.h,
+        ),
+        _menuItem(S.of(context).contactUs, widget.contactUsIcon, () {
+          AlertModule().showContactUsDialog(
+              contactUsBloc: widget.contactUsBloc, context: context);
+        }),
+        SizedBox(
+          height: 10.h,
+        ),
+        _menuItem(S.of(context).faq, widget.faqIcon, () {
+          CustomNavigatorModule.navigatorKey.currentState
+              ?.pushNamed(AppScreenEnum.faq.name);
+        }),
+        SizedBox(
+          height: 10.h,
+        ),
+        _menuItem(S.of(context).usagePolicy, widget.usagePolicyIcon, () {}),
+        if ((SharedPrefModule().userId ?? '').isNotEmpty) ...[
           SizedBox(
             height: 37.h,
           ),
           _menuItem(S.of(context).logout, widget.logoutIcon, () {
             _logout();
           }, isBoldStyle: true),
-          // _moreDesign,
-        ],
-      );
+        ]
+        // _moreDesign,
+      ],
+    );
+  }
 
-  Widget get _imageWithCameraWidget => ShopLogoCameraWidget(
+  Widget _imageWithCameraWidget(
+          {required String mobile,
+          required String name,
+          required String image}) =>
+      ShopLogoCameraWidget(
         placeHolder: widget.shopIcon,
-        shopLogo: '',
+        shopLogo: image,
         cameraIcon: widget.cameraIcon,
         moreBloc: widget.moreBloc,
-        openCamera: widget.openCamera,
-        mobile: '00122234567',
-        name: 'هاجر اسامة',
-        onImagePick: (file) => _updateImage(file),
+        openCameraOrGallery: () => handleCameraOrGallery(),
+        mobile: mobile,
+        name: name,
       );
+
+  void handleCameraOrGallery() {
+    AlertModule().showDialog(
+      context: context,
+      message: S.of(context).selectPhotoFromCameraOrGallery,
+      cancelMessage: S.of(context).gallery,
+      confirmMessage: S.of(context).camera,
+      onCancel: () {
+        _requestGalleryPermission();
+        _listenForGalleryPermission();
+      },
+      onConfirm: () {
+        requestCameraPermission();
+        _listenForCameraPermissionResult();
+      },
+    );
+  }
+
+  void requestCameraPermission() {
+    widget.moreBloc.cameraPermissionBloc
+        .requestPermission(context, Permission.camera);
+    widget.moreBloc.cameraPermissionBloc.listenFormOpenSettings();
+  }
+
+  void _listenForCameraPermissionResult() {
+    widget.moreBloc.cameraPermissionBloc.easyPermissionHandler
+        .isPermissionGrantedStream
+        .listen((event) async {
+      if (event) {
+        XFile? file = await widget.moreBloc.takePhoto();
+        if (file != null) {
+          widget.moreBloc.uploadImage(file.path);
+        }
+      }
+    });
+  }
+
+  void _requestGalleryPermission() {
+    widget.moreBloc.galleryPermissionBloc
+        .requestPermission(context, Permission.photos);
+    widget.moreBloc.galleryPermissionBloc.listenFormOpenSettings();
+  }
+
+  void _listenForGalleryPermission() {
+    widget.moreBloc.galleryPermissionBloc.easyPermissionHandler
+        .isPermissionGrantedStream
+        .listen((event) async {
+      if (event) {
+        XFile? file = await widget.moreBloc.pickFromGallery();
+        if (file != null) {
+          widget.moreBloc.uploadImage(file.path);
+        }
+      }
+    });
+  }
 
   Widget get _logoWidget => Container(
         alignment: Alignment.center,
@@ -254,36 +389,43 @@ class _MoreWidgetState extends BaseState<MoreWidget> {
       );
 */
   Widget _menuItem(String text, String imagePath, VoidCallback onTap,
-          {bool isBoldStyle = false}) =>
-      InkWell(
-        onTap: () => onTap(),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.start,
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: [
-            SizedBox(
-              width: 16.w,
-            ),
-            ImageHelper(
-              image: imagePath,
-              imageType: ImageType.svg,
-              width: 24.w,
-              height: 24.h,
-              color: lightBlackColor,
-              boxFit: BoxFit.fill,
-            ),
-            SizedBox(
-              width: 20.w,
-            ),
-            CustomText(
-                text: text,
-                customTextStyle: isBoldStyle
-                    ? BoldStyle(color: lightBlackColor, fontSize: 20.sp)
-                    : RegularStyle(color: lightBlackColor, fontSize: 16.w)),
-            SizedBox(
-              width: 16.w,
-            ),
-          ],
+          {bool isBoldStyle = false, bool disabled = false}) =>
+      IgnorePointer(
+        ignoring: disabled,
+        child: InkWell(
+          onTap: () => disabled ? null : onTap(),
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.start,
+            crossAxisAlignment: CrossAxisAlignment.center,
+            children: [
+              SizedBox(
+                width: 13.w,
+              ),
+              ImageHelper(
+                image: imagePath,
+                imageType: ImageType.svg,
+                width: 24.w,
+                height: 24.h,
+                color: disabled ? greyColor : lightBlackColor,
+                boxFit: BoxFit.fill,
+              ),
+              SizedBox(
+                width: 16.w,
+              ),
+              CustomText(
+                  text: text,
+                  customTextStyle: isBoldStyle
+                      ? BoldStyle(
+                          color: disabled ? greyColor : lightBlackColor,
+                          fontSize: 20.sp)
+                      : RegularStyle(
+                          color: disabled ? greyColor : lightBlackColor,
+                          fontSize: 16.w)),
+              SizedBox(
+                width: 16.w,
+              ),
+            ],
+          ),
         ),
       );
 
@@ -293,29 +435,39 @@ class _MoreWidgetState extends BaseState<MoreWidget> {
           text: S.of(context).favourites,
           customTextStyle: BoldStyle(fontSize: 20.sp, color: secondaryColor)));
 
-  Widget _accountBalance(String balance) => Padding(
-        padding: EdgeInsets.symmetric(horizontal: 16.w),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: [
-            CustomText(
-                text: S.of(context).accountBalance,
-                customTextStyle:
-                    BoldStyle(fontSize: 20.sp, color: secondaryColor)),
-            Container(
-              decoration: BoxDecoration(
-                  color: redColor, borderRadius: BorderRadius.circular(8.w)),
-              padding: EdgeInsets.symmetric(vertical: 4.h, horizontal: 10.w),
-              child: CustomText(
-                  text: balance,
-                  customTextStyle: RegularStyle(
-                    color: whiteColor,
-                    fontSize: 18.sp,
-                  )),
-            )
-          ],
-        ),
+  StreamBuilder<ApiState<BalanceMapper>> _accountBalance() =>
+      StreamBuilder<ApiState<BalanceMapper>>(
+        stream: widget.moreBloc.balanceStream,
+        initialData: LoadingState(),
+        builder: (context, snapshot) => checkResponseStateWithLoadingWidget(
+            snapshot.data!, context,
+            onSuccess: Padding(
+              padding: EdgeInsets.symmetric(horizontal: 16.w),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                crossAxisAlignment: CrossAxisAlignment.center,
+                children: [
+                  CustomText(
+                      text: S.of(context).accountBalance,
+                      customTextStyle:
+                          BoldStyle(fontSize: 20.sp, color: secondaryColor)),
+                  Container(
+                    decoration: BoxDecoration(
+                        color: redColor,
+                        borderRadius: BorderRadius.circular(8.w)),
+                    padding:
+                        EdgeInsets.symmetric(vertical: 4.h, horizontal: 10.w),
+                    child: CustomText(
+                        text: (snapshot.data?.response?.balance ?? 0.0)
+                            .toString(),
+                        customTextStyle: RegularStyle(
+                          color: whiteColor,
+                          fontSize: 18.sp,
+                        )),
+                  )
+                ],
+              ),
+            )),
       );
 
   void _updateImage(File file) {
